successfully combined and final touch ups
