+++ src/App.css
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 #navbar li{
     margin: 0 0.5rem;
     /* color: #121212; */
@@ -8,9 +7,12 @@
     font-size:small;
     color: white;
     font-weight: 700;
+    
 }
 .App{
 caret-color: yellow;
+background:black;
+/* color:white; */
 
 }
 
@@ -104,9 +106,3 @@ background-color: rgba(76, 79, 82, 0.29);
   ::-webkit-scrollbar-thumb:hover {
     background: #555; 
   }
-=======
-.App{
-    background:black;
-    color:white;
-}
->>>>>>> b429d71730224a1686dd7e76ab3fa8c678445cdf
