+++ src/Redux/Kanhaiya/reducer.js
@@ -1,5 +1,5 @@
 import { Name } from "./action"
 
-export const ReducerK = ( store = {}, action)=>{
+export const ReducerK = ( store = {user:{}, isLogin:false}, action)=>{
 
 }
