+++ src/index.css
@@ -6,7 +6,7 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  /* background-color: black; */
+  /* background-color: black;   */
   box-sizing: border-box;
 }
